feat: use new CAs with commits failed fixed
the "commits failed” field on the CAs previously failed to deduplicate 
commits that failed across days so if commit A failed yesterday and 
today it would count for 2 commits failed, the new CAs fix this so we 
want to consume information from them instead of the old ones.

This commit also reorganizes the timescale aggregation code in the GQL 
API and checks the is_called_from_sentry_app function to decide which 
version of the aggregation to use
