+++ apps/codecov-api/graphql_api/tests/snapshots/analytics_timescale__TestAnalyticsTestCaseNew__gql_query_test_results_timescale__0.json
@@ -6,13 +6,13 @@
           "totalCount": 6,
           "edges": [
             {
-              "cursor": "NC4wfG5hbWU0",
+              "cursor": "Ni4wfG5hbWU0",
               "node": {
                 "name": "name4",
                 "failureRate": 0.0,
                 "flakeRate": 0.0,
-                "avgDuration": 4.0,
-                "totalDuration": 8.0,
+                "avgDuration": 6.0,
+                "totalDuration": 12.0,
                 "totalFailCount": 0,
                 "totalFlakyFailCount": 0,
                 "totalPassCount": 2,
@@ -22,13 +22,13 @@
               }
             },
             {
-              "cursor": "My4wfG5hbWUz",
+              "cursor": "NC41fG5hbWUz",
               "node": {
                 "name": "name3",
                 "failureRate": 1.0,
                 "flakeRate": 0.0,
-                "avgDuration": 3.0,
-                "totalDuration": 6.0,
+                "avgDuration": 4.5,
+                "totalDuration": 9.0,
                 "totalFailCount": 2,
                 "totalFlakyFailCount": 0,
                 "totalPassCount": 0,
@@ -38,13 +38,13 @@
               }
             },
             {
-              "cursor": "Mi4wfG5hbWUy",
+              "cursor": "My4wfG5hbWUy",
               "node": {
                 "name": "name2",
                 "failureRate": 0.0,
                 "flakeRate": 0.0,
-                "avgDuration": 2.0,
-                "totalDuration": 4.0,
+                "avgDuration": 3.0,
+                "totalDuration": 6.0,
                 "totalFailCount": 0,
                 "totalFlakyFailCount": 0,
                 "totalPassCount": 2,
@@ -54,13 +54,13 @@
               }
             },
             {
-              "cursor": "MS4wfG5hbWUx",
+              "cursor": "MS41fG5hbWUx",
               "node": {
                 "name": "name1",
                 "failureRate": 1.0,
                 "flakeRate": 0.0,
-                "avgDuration": 1.0,
-                "totalDuration": 2.0,
+                "avgDuration": 1.5,
+                "totalDuration": 3.0,
                 "totalFailCount": 2,
                 "totalFlakyFailCount": 0,
                 "totalPassCount": 0,
