+++ apps/codecov-api/graphql_api/tests/snapshots/analytics_timescale__TestAnalyticsTestCaseNew__gql_query_test_results_timescale_slowest_parameter__0.json
@@ -6,13 +6,13 @@
           "totalCount": 1,
           "edges": [
             {
-              "cursor": "NC4wfG5hbWU0",
+              "cursor": "Ni4wfG5hbWU0",
               "node": {
                 "name": "name4",
                 "failureRate": 0.0,
                 "flakeRate": 0.0,
-                "avgDuration": 4.0,
-                "totalDuration": 8.0,
+                "avgDuration": 6.0,
+                "totalDuration": 12.0,
                 "totalFailCount": 0,
                 "totalFlakyFailCount": 0,
                 "totalPassCount": 2,
